Added the anchors for spaces tothe proposal page
I did not find much about inputcombo tag so I have added the regular link tag and to be opened in new page so you so notloose the page you are submitting. 
I am not sure if this is ok as we used the whole text as selection area before and now we reduce it to the radio button itself only. 

@autonome Is this a fix for #423 ? Would this work with the overall site @mmmavis?